pyproject.toml
[project]
name = "travel-planner-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.3.0",
    "google-genai>=1.12.1",
    "pytest>=8.3.5",
]
